+++ commands/certificates.go
@@ -5,7 +5,6 @@ import (
 	"bonvoy/envoy"
 	"fmt"
 	"github.com/spf13/cobra"
-	"strings"
 )
 
 type Certificates struct {
@@ -24,7 +23,7 @@ func (r *Registry) Certificates() *Certificates {
 }
 
 func (r *Registry) BuildExpiredCertificatesCommand() *cobra.Command {
-	return &cobra.Command{
+	cmd := &cobra.Command{
 		Use: "expired",
 		Short: "Show all expired certificates",
 		Long:  `Display all expired sidecar certificates as compared to the local Consul agent`,
@@ -34,9 +33,14 @@ func (r *Registry) BuildExpiredCertificatesCommand() *cobra.Command {
 				ServiceName: args[0],
 				Consul: consul.NewClient(),
 			}
-			return controller.Run()
+			restart, err := cmd.Flags().GetBool("restart")
+			if err != nil { return err }
+
+			return controller.Run(restart)
 		},
 	}
+	cmd.Flags().BoolP("restart", "r", false, "If passed, will restart all sidecars that have expired certificates")
+	return cmd
 }
 
 type ExpiredCertificatesController struct {
@@ -44,48 +48,42 @@ type ExpiredCertificatesController struct {
 	Consul consul.Client
 }
 
-func (c *ExpiredCertificatesController) Run() error {
+func (c *ExpiredCertificatesController) Run(restart bool) error {
+	var expiredCerts []envoy.ExpiredCertificate
+	var sidecars []envoy.Instance
+
 	if c.ServiceName != "all" {
 		e, err := envoy.NewFromServiceName(c.ServiceName)
 		if err != nil { return err }
 
-		return c.LookupForSidecar(e)
+		sidecars = append(sidecars, e)
 	} else {
-		sidecars, err := envoy.AllSidecars()
-		if err != nil { return err }
-
-		for _, e := range sidecars {
-			err = c.LookupForSidecar(e)
-			if err != nil { return err }
+		sideResp, err := envoy.AllSidecars()
+		if err != nil {
+			return err
 		}
+
+		sidecars = append(sidecars, sideResp...)
 	}
 
-	return nil
-}
+	for _, e := range sidecars {
+		resp, lErr := e.Certificates().FindExpired()
+		if lErr != nil { return lErr }
 
-func (c *ExpiredCertificatesController) LookupForSidecar(e envoy.Instance) error {
-	data, err := e.Certificates().Get()
-	if err != nil { return err }
+		expiredCerts = append(expiredCerts, resp...)
+	}
 
-	readSerials := map[string]bool{}
+	for _, e := range expiredCerts {
+		fmt.Println(e.ServiceName)
+		fmt.Println("  Envoy Process ID:", e.Pid)
+		fmt.Printf("  Envoy Certificate Expiry: %s (%d days)\n", e.EnvoyExpiration, e.EnvoyDaysUntilExpiration)
+		fmt.Println("  Consul Agent Certificate Expiry: ", e.ConsulExpiration)
 
-	for _, certs := range data.Certificates {
-		for _, cert := range certs.CertificateChain {
-			if readSerials[cert.SerialNumber] {
-				continue // ignore duplicates
-			}
-			readSerials[cert.SerialNumber] = true
-			a := strings.Split(cert.SubjectAltNames[0].Uri, "/")
-			svc := a[len(a)-1]
-			leaf, lErr := c.Consul.Agent().GetConnectLeafCaCertificate(svc)
-			if lErr != nil { return lErr }
+		if restart == true {
+			err := e.Envoy.Restart()
+			if err != nil { return err }
 
-			if cert.ExpirationTime != leaf.ValidBefore {
-				fmt.Println(svc)
-				fmt.Println("  Envoy Process ID:", e.Pid)
-				fmt.Printf("  Envoy Certificate Expiry: %s (%s days)\n", cert.ExpirationTime, cert.DaysUntilExpiration)
-				fmt.Println("  Consul Agent Certificate Expiry: ", leaf.ValidBefore)
-			}
+			fmt.Println("...restarting...Done.")
 		}
 	}